+++ domenetjenester/src/main/java/no/nav/folketrygdloven/kalkulus/domene/entiteter/beregningsgrunnlag/BeregningsgrunnlagPrStatusOgAndelEntitet.java
@@ -61,7 +61,7 @@ public class BeregningsgrunnlagPrStatusOgAndelEntitet extends BaseEntitet {
 
     @Version
     @Column(name = "versjon", nullable = false)
-    private long versjon;
+    private int versjon;
 
     @JsonBackReference
     @ManyToOne(optional = false)
