+++ domenetjenester/src/main/java/no/nav/folketrygdloven/kalkulus/domene/entiteter/beregningsgrunnlag/FaktaAktørEntitet.java
@@ -30,7 +30,7 @@ public class FaktaAktørEntitet extends BaseEntitet {
 
     @Version
     @Column(name = "versjon", nullable = false)
-    private long versjon;
+    private int versjon;
 
     @JsonBackReference
     @OneToOne(optional = false, fetch = FetchType.EAGER)
